consolidating collision code

+++ main/modes/games/bigbug/tilemap_bigbug.c
@@ -824,14 +824,14 @@ void bb_collisionCheck(bb_tilemap_t* tilemap, bb_entity_t* ent, vec_t* previousP
                                          j * BITSHIFT_TILE_SIZE + BITSHIFT_HALF_TILE};
                         // AABB-AABB collision detection begins here
                         // https://tutorialedge.net/gamedev/aabb-collision-detection-tutorial/
-                        if ((int32_t)ent->pos.x - (int32_t)ent->halfWidth
-                                < (int32_t)tilePos.x + (int32_t)BITSHIFT_HALF_TILE
-                            && (int32_t)ent->pos.x + (int32_t)ent->halfWidth
-                                   > (int32_t)tilePos.x - (int32_t)BITSHIFT_HALF_TILE
-                            && (int32_t)ent->pos.y - (int32_t)ent->halfHeight
-                                   < (int32_t)tilePos.y + (int32_t)BITSHIFT_HALF_TILE
-                            && (int32_t)ent->pos.y + (int32_t)ent->halfHeight
-                                   > (int32_t)tilePos.y - (int32_t)BITSHIFT_HALF_TILE)
+                        if (ent->pos.x - (int32_t)ent->halfWidth
+                                < tilePos.x + (int32_t)BITSHIFT_HALF_TILE
+                            && ent->pos.x + (int32_t)ent->halfWidth
+                                   > tilePos.x - (int32_t)BITSHIFT_HALF_TILE
+                            && ent->pos.y - (int32_t)ent->halfHeight
+                                   < tilePos.y + (int32_t)BITSHIFT_HALF_TILE
+                            && ent->pos.y + (int32_t)ent->halfHeight
+                                   > tilePos.y - (int32_t)BITSHIFT_HALF_TILE)
                         {
                             /////////////////////////
                             // Collision detected! //
